refactor: kullna rewrite
The initial Kullna Editor commit.

This changeset rewrites large portions of the codejar repository
and prepares our fork for contributions.

- We maintained CodeJar's history as a nod of appreciation to all the contributors there
- Added linting, formatting, commit message linting
- Added documentation and demonstrations for users
- Added documentation for contributors
- Refactored Cursor and LineNumbers APIs
- Documented all the code
- Cleaned up the external APIs
- Adopted best practices to be compatiable with common syntax highliters by default
- Improved performance of the Gutter
- Added support for Gutter customization
- Made RTL work correctly by default

Notably, we've prepared the gutter APIs such that we can eventually be much smarter
about how they are maintained. Currently, a new element is created in the DOM for
every line number. We would like to change this over time to be an element recycler.